Added a weapon Creator.  Put choice of weapon use in the Fight system.  Working on some UI shit and debugging the program.

+++ CelestialTravels/Character.cs
@@ -11,16 +11,16 @@ namespace CelestialTravels
         public string Race { get; set; }
         public string Job { get; set; }
         public int Credits { get; set; }
-        public int Experience { get; set; }
-        public int CurrentHitPoints { get; set; } = 100;
-        public int TotalHitPoints { get; set; } = 100;
+        public double Experience { get; set; }
+        public double CurrentHitPoints { get; set; } = 100;
+        public double TotalHitPoints { get; set; } = 100;
         public int WeaponSkill { get; set; }
         public int ArmorSkill { get; set; }
         public int NavigationSkill { get; set; }
         public int TimeSkill { get; set; }
-        public int CurrentWeaponDamage { get; set; }
-        public int Roll { get; set; }
-        public int Attack { get; set; }
+        public double CurrentWeaponDamage { get; set; }
+        public double Roll { get; set; }
+        public double Attack { get; set; }
         public int Level { get; set; }
 
     }
